New Stuff:
Cart.html
Main.java: get method ; "/showcart"
ProductController.java: method; renderCart
-------------------------------------------------------
Changes:

CartDao interface: changed getAll() return type (from Map<Product, AtomicLong> to Map<Product, Integer>)
CartDaoMem class: changed getAll() method according to the changed interface

+++ src/main/java/com/codecool/shop/dao/CartDao.java
@@ -11,7 +11,7 @@ public interface CartDao {
 
     void add(Product product);
 
-    Map<Product, AtomicLong> getAll();
+    Map<Product, Integer> getAll();
 
     int getCount();
 }
